[#111] Refactor FoodRequestViewSet and update comments across multiple files
